liking, users, auth all working

+++ properties/views.py
@@ -1,12 +1,15 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from rest_framework.exceptions import NotFound
+from rest_framework.exceptions import NotFound, PermissionDenied
+from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticated
 from .models import Property
 from .serializers.common import PropertySerializer
 from .serializers.populated import PopulatedPropertySerializer
 
 class PropertyListView(APIView):
+
+    permission_classes = (IsAuthenticatedOrReadOnly, )
 
     def get(self, _request):
         properties = Property.objects.all() #querying property from index
@@ -14,6 +17,7 @@ class PropertyListView(APIView):
         return Response(serialized_property.data, status=status.HTTP_200_OK)
 
     def post(self, request):
+        request.data['owner'] = request.user.id
         property_to_create = PropertySerializer(data=request.data)
         if property_to_create.is_valid():
             property_to_create.save()
@@ -21,6 +25,8 @@ class PropertyListView(APIView):
         return Response(property_to_create.errors, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 class PropertyDetailView(APIView):
     """ Controller responsible for get/put/delete requests to /properties/id endpoint """
+
+    permission_classes = (IsAuthenticatedOrReadOnly, )
 
     def get_property(self, pk):
 
@@ -41,13 +47,28 @@ class PropertyDetailView(APIView):
 
     def put(self, request, pk):
         property_to_update = self.get_property(pk=pk)
+        if property_to_update.owner.id != request.user.id:
+            raise PermissionDenied()
         updated_property = PropertySerializer(property_to_update, data=request.data)
         if updated_property.is_valid():
             updated_property.save()
             return Response(updated_property.data, status=status.HTTP_202_ACCEPTED)
         return Response(updated_property.errors, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
-    def delete(self, _request, pk):
+    def delete(self, request, pk):
         property_to_delete = self.get_property(pk=pk)
+        if property_to_delete.owner.id != request.user.id:
+            raise PermissionDenied()
         property_to_delete.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+class PropertyFavoriteView(PropertyDetailView):
+  
+    permission_classes = (IsAuthenticated, )
+
+    def post(self, request, pk):
+        property_to_like = self.get_property(pk=pk)
+        property_to_like.favorited_by.add(request.user.id)
+        property_to_like.save()
+        serialized_liked_property = PopulatedPropertySerializer(property_to_like)
+        return Response(serialized_liked_property.data, status=status.HTTP_201_CREATED)
